Add possibility to specify "extra files" (aka "extras")
It is now possible to tell the package manager that a package needs
certain files installed at specific locations outside of its data
directory. These files are called the "extra files" of a package.
Use cases are things like .desktop files in /usr/share/applications
and the like. To be used with GREAT care! See the "Extra Files"
section in the updated psysmeta(7) section for the details.