Implement dependency checks in RPM fallback backend
Previously, the RPM fallback backend added the dependency
"lsb >= LSBVERSION" to each registered package, but didn't check
if the dependency is actually met. Now, it does both.

Also, fix the return value of the internal find_db_record() function
to be UINT_MAX on error, not -1. It's the same really, but considering
the return type is unsigned int, this seems clearer.