Add DPKG fallback backend (for Debian/Ubuntu and derivatives)
With this backend, the psys library is now really cross-distro and
not just a complex  wrapper around librpm. ;) With RPM and DPKG
covered, psys should have integration support for all currently
LSB-certified distros.

Note that the new backend needs a libdpkg built with position-
independent code (-fPIC), and that the configure switches for
enabling the fallback backends have changed; the updated README
has more.

Together with the new backend comes a bit factoring of code formerly
in the RPM fallback backend to shared locations (libpsys_impl.h or
the new fallback_private.h for fallback-specific things).